% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_summarize_table.R
\name{by_use_case}
\alias{by_use_case}
\title{by_use_case}
\usage{
by_use_case(data, selected_filter)
}
\arguments{
\item{data}{original complete data}

\item{selected_filter}{how you want to filter the table}
}
\value{
end_table
}
\description{
helper function that summarizes the assessment table by use case
}
