% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_summarize_table.R
\name{summarize_table}
\alias{summarize_table}
\title{summarize_table}
\usage{
summarize_table(data, selected_filter)
}
\arguments{
\item{data}{original complete data}

\item{selected_filter}{how you want to filter the table}
}
\value{
summary_table
}
\description{
Creates summary table from the data retrieved from the assessment
table
}
