% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\docType{data}
\name{validStatuses}
\alias{validStatuses}
\title{Valid statuses}
\description{
These status strings correspond to colors as defined in Bootstrap's CSS.
Although the colors can vary depending on the particular CSS selector, they
generally appear as follows:
}
\details{
`r generate_color_documentation(c("primary", "secondary", "success", "info", "warning", "danger"))`
}
\keyword{internal}
